feat: protocol support for V11 fetch w/ rackID
These are the initial protocol beginnings to add support for follower
fetching (KIP-392). This commit simply adds configuration support for a
consumer to specify its rack identifier and to flow it in a V11
FetchRequest and correctly decode the "preferredReplica" field in the
response. Currently no further action is taken. A future PR will be
required to act upon this response.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>