fix: use modern kafka server.properties in vagrant
Rather than copying and templating an old copy of a Kafka
server.properties, modify the install_cluster.sh to append a small
number of custom values to the end of the Java properties file as
shipped with each Kafka version. This ensures we are otherwise testing
the defaults for each version apart from where we have made
customisations.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>